third_party/blink/web_tests/html/text_level_semantics/mark-element-expected.txt

Various tests for the mark element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

<p> closes <mark>:
PASS paragraph1.parentNode.nodeName == "mark" is false
<b> does not close <mark>:
PASS b1.parentNode.nodeName is "MARK"
Residual style:
PASS getWeight("mark2") is "700"
PASS getWeight("span1") is "700"
PASS successfullyParsed is true

TEST COMPLETE
